requirements.txt
# Core
python-binance==1.0.17
polars==0.20.3
pandas==2.1.3
numpy==1.26.2

# Deep Learning
torch==2.2.0
pytorch-lightning==2.2.0

# Trading & Backtesting
vectorbt==0.27.3

# Data & Persistence
pyarrow==14.0.1

# Configuration & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0

# API & Logging
fastapi==0.104.1
uvicorn==0.24.0
structlog==23.2.0

# Testing & Dev
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
